+++ src/assets/routes.ts
@@ -1,9 +1,8 @@
-export const ROUTES_INFO = [
-  '/terms',
-  '/privacy',
-  '/shipping',
-  '/faq',
-  '/returns',
-];
+export const ROUTE_HOME = '/';
 export const ROUTE_CONTACT = '/contact';
 export const ROUTE_ABOUT = '/about';
+
+export const ROUTE_SHOP = 'https://shop.enyo.gg';
+
+export const ROUTE_TERMS = `${ROUTE_SHOP}/policies/terms-of-service`;
+export const ROUTE_PRIVACY = `${ROUTE_SHOP}/policies/privacy-policy`;