+++ src/components/NavList.tsx
@@ -1,17 +1,18 @@
-import React, { FunctionComponent } from 'react';
-import { makeStyles, Theme } from '@material-ui/core/styles';
 import {
-  Typography,
   List,
-  ListItemText,
   ListItem,
-  Link,
+  ListItemText,
+  Typography,
+  useMediaQuery,
 } from '@material-ui/core';
+import { makeStyles, useTheme } from '@material-ui/core/styles';
 import clsx from 'clsx';
+import React, { FunctionComponent } from 'react';
 import { Link as RouterLink } from 'react-router-dom';
-import { LinkedItem } from '../types';
-import WingedBorder from './WingedBorder';
 import { BREAKPOINT_MOBILE } from '../assets/consts';
+import { LinkedItem } from '../types';
+import Expand from './Expand';
+import WingedBorder from './WingedBorder';
 
 interface NavListProps {
   title: string;
@@ -20,33 +21,24 @@ interface NavListProps {
 }
 
 const useStyles = makeStyles(
-  (theme: Theme) => ({
+  () => ({
     root: {
       display: 'inline-flex',
       flexDirection: 'column',
       width: '100%',
     },
     title: {
-      paddingLeft: theme.spacing(2),
+      fontSize: '1.6rem',
     },
     wing: {
-      marginTop: '-0.8rem',
+      marginBottom: '-0.6rem',
     },
     listText: {
       marginTop: 0,
       marginBottom: 0,
     },
-    listLink: {
-      [theme.breakpoints.down(BREAKPOINT_MOBILE)]: {
-        fontSize: '0.9rem',
-      },
-    },
-    listItem: {
-      paddingRight: 0,
-      [theme.breakpoints.down(BREAKPOINT_MOBILE)]: {
-        paddingTop: 0,
-        paddingBottom: 0,
-      },
+    text: {
+      fontSize: '0.9rem',
     },
   }),
   {
@@ -60,28 +52,48 @@ const NavList: FunctionComponent<NavListProps> = ({
   className,
 }) => {
   const classes = useStyles();
+  const theme = useTheme();
+  const isMobile = useMediaQuery(theme.breakpoints.down(BREAKPOINT_MOBILE));
+
+  const navList = (
+    <List dense component="nav">
+      {items.map((item) => {
+        const isLink = item.link.startsWith('http');
+        return (
+          <ListItem
+            button
+            key={item.text}
+            component={isLink ? 'a' : RouterLink}
+            to={isLink ? undefined : item.link}
+            href={isLink ? item.link : undefined}
+            disableGutters
+          >
+            <ListItemText className={classes.listText} disableTypography>
+              <Typography
+                variant="body2"
+                className={classes.text}
+                component="span"
+              >
+                {item.text}
+              </Typography>
+            </ListItemText>
+          </ListItem>
+        );
+      })}
+    </List>
+  );
+
+  if (isMobile) {
+    return <Expand title={title}>{navList}</Expand>;
+  }
 
   return (
     <div className={clsx(classes.root, className)}>
       <Typography variant="h5" className={classes.title}>
         {title}
       </Typography>
-      <WingedBorder right direction="up" className={classes.wing} />
-      <List dense>
-        {items.map((item) => (
-          <ListItem key={item.text} className={classes.listItem}>
-            <ListItemText className={classes.listText}>
-              <Link
-                component={RouterLink}
-                to={item.link}
-                className={classes.listLink}
-              >
-                {item.text}
-              </Link>
-            </ListItemText>
-          </ListItem>
-        ))}
-      </List>
+      <WingedBorder right direction="down" className={classes.wing} />
+      {navList}
     </div>
   );
 };