+++ src/pages/Contact.tsx
@@ -25,7 +25,9 @@ interface FormElements {
   name: string;
   email: string;
   message: string;
-  website: string; // Honeypot
+  team: string;
+  website: string;
+  password: string; // Honeypot
 }
 
 enum FormState {
@@ -75,6 +77,8 @@ const Contact: FunctionComponent = () => {
     name: sessionStorage.getItem('contact.name') || undefined,
     email: sessionStorage.getItem('contact.email') || undefined,
     message: sessionStorage.getItem('contact.message') || undefined,
+    team: sessionStorage.getItem('contact.team') || undefined,
+    website: sessionStorage.getItem('contact.website') || undefined,
   };
 
   const hasFilledForm =
@@ -84,7 +88,9 @@ const Contact: FunctionComponent = () => {
     name: '',
     email: '',
     message: '',
+    team: '',
     website: '',
+    password: '',
   });
   const [apiErrorText, setApiErrorText] = useState('');
   const [formState, setFormState] = useState<FormState>(
@@ -97,7 +103,7 @@ const Contact: FunctionComponent = () => {
     if (!hasFilledForm) {
       setSending(true);
 
-      if (values.website !== '') {
+      if (values.password !== '') {
         // Honeypot capture
         setApiErrorText(t('pages.contact.form.honey'));
         setFormState(FormState.error);
@@ -112,6 +118,8 @@ const Contact: FunctionComponent = () => {
             sessionStorage.setItem('contact.name', values.name);
             sessionStorage.setItem('contact.email', values.email);
             sessionStorage.setItem('contact.message', values.message);
+            sessionStorage.setItem('contact.team', values.team);
+            sessionStorage.setItem('contact.website', values.website);
             setFormState(FormState.sent);
             setSending(false);
           })
@@ -173,7 +181,7 @@ const Contact: FunctionComponent = () => {
         className={classes.form}
       >
         <Grid container spacing={2}>
-          <Grid item xs={12} sm>
+          <Grid item xs={12} sm={6}>
             <TextField
               name="name"
               label={t('pages.contact.labels.name')}
@@ -184,10 +192,10 @@ const Contact: FunctionComponent = () => {
               required
               disabled={disableControls}
               className={classes.width}
-              value={storedValues.name}
+              defaultValue={storedValues.name}
             />
           </Grid>
-          <Grid item xs={12} sm>
+          <Grid item xs={12} sm={6}>
             <TextField
               name="email"
               label={t('pages.contact.labels.email')}
@@ -198,7 +206,33 @@ const Contact: FunctionComponent = () => {
               required
               disabled={disableControls}
               className={classes.width}
-              value={storedValues.email}
+              defaultValue={storedValues.email}
+            />
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <TextField
+              name="team"
+              label={t('pages.contact.labels.team')}
+              variant="outlined"
+              type="text"
+              error={formState === FormState.error}
+              onChange={(e) => handleChange(e)}
+              disabled={disableControls}
+              className={classes.width}
+              defaultValue={storedValues.team}
+            />
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <TextField
+              name="website"
+              label={t('pages.contact.labels.website')}
+              variant="outlined"
+              type="text"
+              error={formState === FormState.error}
+              onChange={(e) => handleChange(e)}
+              disabled={disableControls}
+              className={classes.width}
+              defaultValue={storedValues.website}
             />
           </Grid>
           <Grid item xs={12}>
@@ -216,11 +250,11 @@ const Contact: FunctionComponent = () => {
               required
               disabled={disableControls}
               className={classes.width}
-              value={storedValues.message}
+              defaultValue={storedValues.message}
             />
             <input
               type="text"
-              name="website"
+              name="password"
               tabIndex={-1}
               autoComplete="off"
               onChange={(e) => handleChange(e)}