+++ src/pages/info/types.ts
@@ -1,7 +1,0 @@
-export interface InfoPage {
-  title: string;
-  updated: Date;
-  translations: {
-    en: string;
-  };
-}