+++ src/types.ts
@@ -1,7 +1,6 @@
 export interface LinkedItem {
   text: string;
   link: string;
-  fallbackIcon?: JSX.Element;
 }
 
 export interface KeyValue {